Move CTHandler
